StopAtEOF: keep sending lines until EOF
When a StopAtEOF() is called the code should continue to send all lines
to the Lines channel. The issue here is if the caller is not ready to
receive a new line the code blocks as it is using a unbuffered channel.
However <-tail.Dying() would return in this case so the line was
skipped. This means that the caller did not get all lines until EOF.
Now we still want to skip in case any other reason for kill was given
therefore add special logic to only not read the Dying channel on the
EOF case.

The one downside is that StopAtEOF() could block forever if the caller
never reads new Lines but this seems logical to me. If the caller wants
to wait for EOF but never reads remaining Lines this would be a bug on
their end.

Fixes #37

Signed-off-by: Paul Holzinger <pholzing@redhat.com>